+++ halotools/mock_observables/isolation_functions/engines/__init__.py
@@ -1,0 +1,6 @@
+# Licensed under a 3-clause BSD style license - see LICENSE.rst
+
+from __future__ import absolute_import, division, print_function, unicode_literals
+
+from .spherical_isolation_engine import spherical_isolation_engine
+from .cylindrical_isolation_engine import cylindrical_isolation_engine
