+++ halotools/mock_observables/pair_counters/cpairs/__init__.py
@@ -1,6 +1,6 @@
 # Licensed under a 3-clause BSD style license - see LICENSE.rst
 
-from __future__ import (absolute_import, division, print_function, unicode_literals)
+from __future__ import absolute_import, division, print_function, unicode_literals
 
 from .pairwise_distances import *
 from .npairs_3d_engine import npairs_3d_engine
@@ -9,7 +9,5 @@ from .npairs_xy_z_engine import npairs_xy_z_engine
 from .npairs_jackknife_3d_engine import npairs_jackknife_3d_engine
 from .npairs_s_mu_engine import npairs_s_mu_engine
 from .npairs_per_object_3d_engine import npairs_per_object_3d_engine
-from .spherical_isolation_engine import spherical_isolation_engine
-from .cylindrical_isolation_engine import cylindrical_isolation_engine
 from .pairwise_distance_3d_engine import pairwise_distance_3d_engine
 from .pairwise_distance_xy_z_engine import pairwise_distance_xy_z_engine
